Types of tests & BDD (behaviour driven development)
The BDD approach can largely be divided into two main parts. The first is the practice of using examples written in ubiquitous language to illustrate behaviours (how users will interact with the product).

The second part is the practice of using those examples as the basis of automated tests. As well as checking functionality for the user, this ensures the system works as defined by the business throughout the project lifetime.

// Given ( prepare the objects in the environment )

//When ( use the objects with a specific behaviour )

//Then (with the results provided in the *When* see if it's the same in the *Then* )